% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPhenotype.R
\name{computePerformanceMetricsFromCounts}
\alias{computePerformanceMetricsFromCounts}
\title{Compute performance metrics based on a 2-by-2 counts table}
\usage{
computePerformanceMetricsFromCounts(counts)
}
\arguments{
\item{counts}{Counts as created by the \code{\link{testPhenotypeAlgorithm}} function.}
}
\value{
A tibble with the statistics metrics added to the counts table.
}
\description{
Compute performance metrics based on a 2-by-2 counts table
}
\examples{
counts <- tibble::tibble(
  truePositives = 120,
  trueNegatives = 800,
  falsePositives = 30,
  falseNegatives = 50,
  cutPoint = "EV"
)
computePerformanceMetricsFromCounts(counts)

}
