% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCovariateSettings.R
\name{createDefaultAcuteCovariateSettings}
\alias{createDefaultAcuteCovariateSettings}
\title{Create the default covariate settings}
\usage{
createDefaultAcuteCovariateSettings(
  excludedCovariateConceptIds = c(),
  includedCovariateIds = c(),
  addDescendantsToExclude = FALSE,
  startDays = 0,
  endDays = 7
)
}
\arguments{
\item{excludedCovariateConceptIds}{A list of conceptIds to exclude from featureExtraction.  These should include all
concept_ids that were used to define the xSpec model (default=NULL)}

\item{includedCovariateIds}{A list of covariate IDs that should be restricted to.}

\item{addDescendantsToExclude}{Should descendants of excluded concepts also be excluded? (default=FALSE)}

\item{startDays}{The days to include prior to the cohort start date (default=0)}

\item{endDays}{The days to include after the cohort start date (default=7)}
}
\description{
Create the default covariate settings for acute diseases
}
\details{
Function to create the default covariate settings for acute diseases
}
