% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePhenoModel.R
\name{createPhenoModel}
\alias{createPhenoModel}
\title{createPhenoModel}
\usage{
createPhenoModel(connectionDetails = list(), xSpecCohort = "",
  cdmDatabaseSchema = "", cohortDatabaseSchema = "",
  cohortDatabaseTable = "", outDatabaseSchema = "",
  trainOutFile = "", exclCohort = "", estPPV = 1,
  modelAnalysisId = "1", excludedConcepts = c(),
  cdmShortName = "CDM", mainPopnCohort = 0, lowerAgeLimit = 0,
  upperAgeLimit = 120, gender = c(8507, 8532),
  startDate = "19000101", endDate = "21000101")
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function createConnectionDetails
in the DatabaseConnector package.}

\item{xSpecCohort}{The number of the "extremely specific (xSpec)" cohort definition id in
the cohort table (for noisy positives)}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this database. On SQL Server, this should
specifiy both the database and the schema, so for example
'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the cohort
data used to define the at risk cohort is available. Requires read
permissions to this database.}

\item{cohortDatabaseTable}{The tablename that contains the at risk cohort. The expectation is
cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outDatabaseSchema}{The name of a database schema where the user has write capability.  A
temporary cohort table will be created here.}

\item{trainOutFile}{A string designation for the training model file}

\item{exclCohort}{The number of the "extremely sensitive (xSens)" cohort definition id
in the cohort table (used to estimate population prevalence and to
exclude subjects from the noisy positives)}

\item{estPPV}{A value between 0 and 1 as an estimate for the positive predictive
value for the exclCohort}

\item{modelAnalysisId}{Another string for designating the name for the model files}

\item{excludedConcepts}{A list of conceptIds to exclude from featureExtraction}

\item{cdmShortName}{A short name for the current database (CDM)}

\item{mainPopnCohort}{The number of the cohort to be used as a base population for the model
(default=NULL)}

\item{lowerAgeLimit}{The lower age for subjects in the model (default=NULL)}

\item{upperAgeLimit}{The upper age for subjects in the model (default=NULL)}

\item{gender}{The gender(s) to be included (default c(8507, 8532))}

\item{startDate}{The starting date for including subjects in the model (default=NULL)}

\item{endDate}{The ending date for including subjects in the model (default=NULL)}
}
\description{
Create the phenotype model
}
\details{
Function to create a diagnostic prediction model for a health outcome of interest using the xSpec
cohort.  The model may be applied to the evaluation cohort to determine probabilities for each
subject for the health outcome of interest.
}
