% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePhenoModel.R
\name{createPhenotypeModel}
\alias{createPhenotypeModel}
\title{Create the phenotype model}
\usage{
createPhenotypeModel(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outDatabaseSchema,
  modelOutputFileName = "train",
  xSpecCohortId,
  xSensCohortId,
  prevalenceCohortId = xSensCohortId,
  covariateSettings,
  mainPopulationCohortId = 0,
  mainPopulationCohortIdStartDay = 0,
  mainPopulationCohortIdEndDay = 0,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  visitLength = 3,
  gender = c(8507, 8532),
  startDate = "19000101",
  endDate = "21000101",
  removeSubjectsWithFutureDates = TRUE,
  cdmVersion = "5",
  outFolder = getwd(),
  modelType = "chronic"
)
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function createConnectionDetails
in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this database. On SQL Server, this should
specifiy both the database and the schema, so for example
'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the cohort
data used to define the at risk cohort is available. Requires read
permissions to this database.}

\item{cohortTable}{The tablename that contains the at risk cohort. The expectation is
cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outDatabaseSchema}{The name of a database schema where the user has write capability.  A
temporary cohort table will be created here.}

\item{modelOutputFileName}{A string designation for the training model file}

\item{xSpecCohortId}{The number of the "extremely specific (xSpec)" cohort definition id in
the cohort table (for noisy positives)}

\item{xSensCohortId}{The number of the "extremely sensitive (xSens)" cohort definition id
in the cohort table (used to exclude subjects from the base population)}

\item{prevalenceCohortId}{The number of the cohort definition id to determine the disease prevalence,
(default=xSensCohortId)}

\item{covariateSettings}{A covariateSettings object as generated using createCovariateSettings()}

\item{mainPopulationCohortId}{The number of the cohort ID to be used as a base population for the model
(default=NULL)}

\item{mainPopulationCohortIdStartDay}{The number of days relative to the mainPopulationCohortId cohort start date
to begin including visits (default=0)}

\item{mainPopulationCohortIdEndDay}{The number of days relative to the mainPopulationCohortId cohort start date
to end including visits (default=0)}

\item{lowerAgeLimit}{The lower age for subjects in the model (default=NULL)}

\item{upperAgeLimit}{The upper age for subjects in the model (default=NULL)}

\item{visitLength}{The minimum length of index visit for noisy negative comparison (default=3)}

\item{gender}{The gender(s) to be included (default c(8507, 8532))}

\item{startDate}{The starting date for including subjects in the model (default=NULL)}

\item{endDate}{The ending date for including subjects in the model (default=NULL)}

\item{removeSubjectsWithFutureDates}{Should dates be checked to remove future dates (default=TRUE)}

\item{cdmVersion}{The CDM version of the database (default=5)}

\item{outFolder}{The folder where the output files will be written (default=working directory)}

\item{modelType}{The type of health outcome in the model either "acute" or "chronic" (Default = "chronic")}
}
\description{
Create the phenotype model
}
\details{
Function to create a diagnostic prediction model for a health outcome of interest using the xSpec
cohort.  The model may be applied to the evaluation cohort to determine probabilities for each
subject for the health outcome of interest.
}
