% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createPheValuatorAnalysis}
\alias{createPheValuatorAnalysis}
\title{Create a PheValuator analysis specification}
\usage{
createPheValuatorAnalysis(
  analysisId,
  description,
  createEvaluationCohortArgs,
  testPhenotypeAlgorithmArgs
)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific
set of analysis choices.}

\item{description}{A short description of the analysis.}

\item{createEvaluationCohortArgs}{An object representing the arguments to be used when calling
the \code{\link{createCreateEvaluationCohortArgs}} function.}

\item{testPhenotypeAlgorithmArgs}{Should the \code{\link{createTestPhenotypeAlgorithmArgs}} function be used in this
analysis?}
}
\value{
A list object of class \code{"pheValuatorAnalysis"} containing the analysis specification.
  Includes the analysis ID, description, and argument specifications for creating the evaluation cohort
  and testing the phenotype algorithm.
}
\description{
Create a PheValuator analysis specification
}
\details{
Create a set of analysis choices, to be used with the \code{\link{runPheValuatorAnalyses}} function.
}
\examples{
evalCohortArgs <- createCreateEvaluationCohortArgs(
  xSpecCohortId = 1,
  xSensCohortId = 2,
  prevalenceCohortId = 3
)
testArgs <- createTestPhenotypeAlgorithmArgs(
  phenotypeCohortId = 10
)
analysis <- createPheValuatorAnalysis(
  analysisId = 1,
  description = "Main analysis",
  createEvaluationCohortArgs = evalCohortArgs,
  testPhenotypeAlgorithmArgs = testArgs
)

}
