% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{exportPheValuatorResults}
\alias{exportPheValuatorResults}
\title{Export the PheValuator results to CSV files}
\usage{
exportPheValuatorResults(
  outputFolder,
  exportFolder = file.path(outputFolder, "export"),
  pheValuatorAnalysisList,
  packageName = NULL,
  cohortToCreateFile = "settings/CohortsToCreate.csv",
  baseUrl = NULL,
  cohortSetReference = NULL,
  minCellCount = 5
)
}
\arguments{
\item{outputFolder}{Name of the folder where all the outputs will be written to.}

\item{exportFolder}{Name of the folder where all the exported files will be written to.}

\item{pheValuatorAnalysisList}{A list of objects of type \code{pheValuatorAnalysis} as created using
the \code{\link{createPheValuatorAnalysis}} function.}

\item{packageName}{The name of the package containing the cohort definitions. Can be left NULL if
\code{baseUrl} and \code{cohortSetReference} have been specified.}

\item{cohortToCreateFile}{The location of the cohortToCreate file within the package. Is ignored if \code{cohortSetReference} is specified.}

\item{baseUrl}{The base URL for the WebApi instance, for example:
"http://server.org:80/WebAPI". Can be left NULL if no cohort meta-data needs to be exported.}

\item{cohortSetReference}{A data frame with four columns, as described in the details. Can be left NULL if
\code{packageName} and \code{cohortToCreateFile} have been specified.}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}
}
\description{
Export the PheValuator results to CSV files
}
\details{
Currently three ways of executing this function are supported, which determines how the meta-data on the cohorts is retrieved:

\enumerate{
  \item Embedded in a study package, assuming the cohort definitions are stored in that package using the \code{ROhdsiWebApi::insertCohortDefinitionSetInPackage}. In this case, the \code{packageName} and \code{cohortToCreateFile} should be populated.
  \item Using a WebApi interface to retrieve the cohort definitions. In this case, the \code{baseUrl} and \code{cohortSetReference} should be populated.
  \item Skipping meta-data on the cohorts altogether. In this case, only the \code{cohortSetReference} needs to be populated.
}

The \code{cohortSetReference} argument must be a data frame with  the following columns:
\describe{
\item{atlasId}{The cohort ID in ATLAS.}
\item{atlasName}{The full name of the cohort. This will be shown in the Shiny app.}
\item{cohortId}{The cohort ID to use in the package. Usually the same as the cohort ID in ATLAS.}
\item{name}{A short name for the cohort, to use to create file names. do not use special characters.}
}
}
