% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPerformanceCharacteristics.R
\name{getPerformanceCharacteristics}
\alias{getPerformanceCharacteristics}
\title{testPhenotype}
\usage{
getPerformanceCharacteristics(modelFileName = "", cdmShortName = "")
}
\arguments{
\item{modelFileName}{The full file name with path for the model file}

\item{cdmShortName}{A short name for the current database (CDM)}
}
\value{
A list containg 2 dataframes: 1) dataframe with the model performance characteristics
                              2) dataframe with top 10 positive predictors
}
\description{
Get Model Performance Characteristics
}
\details{
This function will extract the model performance characteristics (e.g., AUC, calibration data) from
the .rds output file from the model
}
