% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchPheValuatorExplorer}
\alias{launchPheValuatorExplorer}
\title{Launch the PheValuator Explorer Shiny app}
\usage{
launchPheValuatorExplorer(dataFolder, launch.browser = FALSE)
}
\arguments{
\item{dataFolder}{A folder where the prepared PheValuator results are stored. Use
the \code{\link{prepareForShiny}} function to generate this data file.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the PheValuator Explorer Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the PheValuator results.
}
