% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{prepareForShiny}
\alias{prepareForShiny}
\title{Preprare results for Shiny viewer}
\usage{
prepareForShiny(dataFolder)
}
\arguments{
\item{dataFolder}{folder where the exported zip files are stored. Use
the \code{\link{exportPheValuatorResults}} function to generate these zip files.
Zip files containing results from multiple databases can be placed in the same
folder.}
}
\description{
Processes one or more zip files generated by the \code{\link{exportPheValuatorResults}} function, so they
can be viewed in a Shiny app.
}
