% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputFunctions.R
\name{prettyPhenotypeOutput}
\alias{prettyPhenotypeOutput}
\title{Table ready summary of  results from PheValuator analyses}
\usage{
prettyPhenotypeOutput(outputFolder, label = NULL)
}
\arguments{
\item{outputFolder}{Name of the folder where all the outputs was written to..}

\item{label}{Optional string for the the Data Set column value.}
}
\value{
A data frame of results.
}
\description{
Table ready summary of  results from PheValuator analyses
}
