% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecalibrateModel.R
\name{recalibrateModel}
\alias{recalibrateModel}
\title{Create the evaluation cohort}
\usage{
recalibrateModel(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  prevalenceCohortId,
  mainPopulationCohortId = 0,
  mainPopulationCohortIdStartDay = 0,
  mainPopulationCohortIdEndDay = 0,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  gender = c(8507, 8532),
  startDate = "19000101",
  endDate = "21000101",
  cdmVersion = "5",
  modelId = "main",
  evaluationCohortId = "main",
  modelFileDirectory = NULL,
  destinationModelFileDirectory = NULL,
  evalFileDirectory = NULL,
  destinationevalFileDirectory = NULL,
  removeSubjectsWithFutureDates = TRUE
)
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function
createConnectionDetails in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where
the cohort data used to define the at risk cohort is
available. Requires read permissions to this database.}

\item{cohortTable}{The tablename that contains the at risk cohort. The
expectation is cohortTable has format of COHORT table:
cohort_concept_id, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{prevalenceCohortId}{The number of the cohort definition id to determine the
disease prevalence.}

\item{mainPopulationCohortId}{The number of the cohort to be used as a base population for
the model. If set to 0, the entire database population will be
used.}

\item{mainPopulationCohortIdStartDay}{The number of days relative to the mainPopulationCohortId
cohort start date to begin including visits.}

\item{mainPopulationCohortIdEndDay}{The number of days relative to the mainPopulationCohortId
cohort start date to end including visits.}

\item{lowerAgeLimit}{The lower age for subjects in the model.}

\item{upperAgeLimit}{The upper age for subjects in the model.}

\item{gender}{The gender(s) to be included.}

\item{startDate}{The starting date for including subjects in the model.}

\item{endDate}{The ending date for including subjects in the model.}

\item{cdmVersion}{The CDM version of the database.}

\item{modelId}{A string used to generate the file names for this model.}

\item{evaluationCohortId}{A string used to generate the file names for this evaluation cohort.}

\item{modelFileDirectory}{Directory name of the RDS model file to recalibrate}

\item{destinationModelFileDirectory}{Directory name to put the recalibrated model file}

\item{evalFileDirectory}{Directory name of the RDS evaluation file to get cohort}

\item{destinationevalFileDirectory}{Directory name to put the cohort RDS file}

\item{removeSubjectsWithFutureDates}{For buggy data with data in the future: ignore subjects with
dates in the future?}
}
\description{
Recalibrates a PheValuator model file
}
\details{
Recalibrates a PheValuator model file based on a the prevalence calculated by the prevalence cohort
}
