% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runPheValuatorAnalyses}
\alias{runPheValuatorAnalyses}
\title{Run a list of analyses}
\usage{
runPheValuatorAnalyses(
  phenotype,
  cohortDefinitionSet = data.frame(),
  analysisName = "Main",
  runDateTime = format(Sys.time(), "\%b \%d \%Y \%X"),
  connectionDetails,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  workDatabaseSchema = cdmDatabaseSchema,
  databaseId = cdmDatabaseSchema,
  cdmVersion = 5,
  outputFolder,
  priorModelToUse = NULL,
  pheValuatorAnalysisList
)
}
\arguments{
\item{phenotype}{Name of the phenotype for analysis}

\item{cohortDefinitionSet}{Data.frame of cohorts must include columns cohortId, cohortName, json, sql - should include all
cohort definitions needed to replicate PheValuator analysis}

\item{analysisName}{Name of the analysis}

\item{runDateTime}{Starting date and time of the PheValuator run}

\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{oracleTempSchema}{DEPRECATED: use \code{tempEmulationSchema} instead.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where
the cohort data used to define the at risk cohort is
available. Requires read permissions to this database.}

\item{cohortTable}{The tablename that contains the at risk cohort. The
expectation is cohortTable has format of COHORT table:
cohort_concept_id, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{workDatabaseSchema}{The name of the database schema that is the location where
a table can be created and afterwards removed.
Requires write permissions to this database.}

\item{databaseId}{Name of the database in the analysis}

\item{cdmVersion}{Define the OMOP CDM version used: currently supports "5".}

\item{outputFolder}{Name of the folder where all the outputs will be written to.}

\item{priorModelToUse}{folder where a previously developed model to use in analysis will be found}

\item{pheValuatorAnalysisList}{A list of objects of type \code{pheValuatorAnalysis} as created using
the \code{\link{createPheValuatorAnalysis}} function.}
}
\value{
A data frame specifying where the constructed evaluation cohort and phenotype evaluation results can be found
in the local file system.
}
\description{
Run a list of analyses
}
\details{
Run a list of analyses.
}
\examples{
\dontrun{
connectionDetails <- DatabaseConnector::createConnectionDetails(
  dbms = "postgresql",
  server = "myserver",
  user = "user",
  password = "password"
)
analysisList <- list(
  createPheValuatorAnalysis(
    analysisId = 1,
    description = "Main analysis",
    createEvaluationCohortArgs = createCreateEvaluationCohortArgs(
      xSpecCohortId = 1,
      prevalenceCohortId = 3
    ),
    testPhenotypeAlgorithmArgs = createTestPhenotypeAlgorithmArgs(
      phenotypeCohortId = 10
    )
  )
)
runPheValuatorAnalyses(
  phenotype = "Diabetes",
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = "cdm.dbo",
  outputFolder = tempdir(),
  pheValuatorAnalysisList = analysisList
)
}

}
