% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{savePheValuatorAnalysisList}
\alias{savePheValuatorAnalysisList}
\title{Save a list of phevaluatorAnalysis to file}
\usage{
savePheValuatorAnalysisList(pheValuatorAnalysisList, file)
}
\arguments{
\item{pheValuatorAnalysisList}{The pheValuatorAnalysis list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\value{
No return value, called for side effects. Writes a JSON file containing the analysis list
  to the specified file path.
}
\description{
Write a list of objects of type \code{pheValuatorAnalysis} to file. The file is in JSON format.
}
\examples{
\dontrun{
# Requires pre-existing analysis list object; example usage:
# savePheValuatorAnalysisList(pheValuatorAnalysisList = myAnalysisList,
#                             file = file.path(tempdir(), "analysisList.json"))
}

}
