% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{summarizePheValuatorAnalyses}
\alias{summarizePheValuatorAnalyses}
\title{Summarize results of PheValuator analyses}
\usage{
summarizePheValuatorAnalyses(referenceTable, outputFolder)
}
\arguments{
\item{referenceTable}{A reference table as created using \code{\link{runPheValuatorAnalyses}}.}

\item{outputFolder}{The output folder used when calling \code{\link{runPheValuatorAnalyses}}.}
}
\value{
A data frame of results.
}
\description{
Summarize results of PheValuator analyses
}
\examples{
\dontrun{
refTable <- runPheValuatorAnalyses(
  phenotype = "Diabetes",
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = "cdm.dbo",
  outputFolder = tempdir(),
  pheValuatorAnalysisList = analysisList
)
results <- summarizePheValuatorAnalyses(refTable, outputFolder = tempdir())
}

}
