% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestConcepts.R
\name{testConcepts}
\alias{testConcepts}
\title{testConcepts}
\usage{
testConcepts(connectionDetails = list(), xSpecCohort = "",
  cdmDatabaseSchema = "", cohortDatabaseSchema = "",
  cohortDatabaseTable = "", outDatabaseSchema = "", testOutFile = "",
  exclCohort = "", modelAnalysisId = "1", inclConcepts = c(),
  cdmShortName = "CDM", mainPopnCohort = 0, lowerAgeLimit = 0,
  upperAgeLimit = 120, gender = c(8507, 8532),
  startDate = "19000101", endDate = "21000101")
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function createConnectionDetails
in the DatabaseConnector package.}

\item{xSpecCohort}{The number of the "extremely specific (xSpec)" cohort definition id in
the cohort table (for noisy positives)}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this database. On SQL Server, this should
specifiy both the database and the schema, so for example
'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the cohort
data used to define the at risk cohort is available. Requires read
permissions to this database.}

\item{cohortDatabaseTable}{The tablename that contains the at risk cohort. The expectation is
cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outDatabaseSchema}{The name of a database schema where the user has write capability.  A
temporary cohort table will be created here.}

\item{testOutFile}{A string designation for the test model file}

\item{exclCohort}{The number of the "extremely sensitive (xSens)" cohort definition id
in the cohort table (used to estimate population prevalence and to
exclude subjects from the noisy positives)}

\item{modelAnalysisId}{Another string for designating the name for the model files}

\item{inclConcepts}{A list of concepts that were included in the definition}

\item{cdmShortName}{A short name for the current database (CDM)}

\item{mainPopnCohort}{The number of the cohort to be used as a base population for the model
(default=NULL)}

\item{lowerAgeLimit}{The lower age for subjects in the model (default=NULL)}

\item{upperAgeLimit}{The upper age for subjects in the model (default=NULL)}

\item{gender}{The gender(s) to be included (default c(8507, 8532))}

\item{startDate}{The starting date for including subjects in the model (default=NULL)}

\item{endDate}{The ending date for including subjects in the model (default=NULL)}
}
\description{
Test the concept sets used in the models
}
\details{
Function to examine the concepts in the xSpec cohort compared to the noisy negative cohort.
This will give a sense of the comrrectnes/completeness of the concept set used for the model
}
