% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPhenotype.R
\name{testPhenotype}
\alias{testPhenotype}
\title{testPhenotype}
\usage{
testPhenotype(connectionDetails = list(), cutPoints = c(0.1, 0.2, 0.3,
  0.4, 0.5, "EV", 0.6, 0.7, 0.8, 0.9), resultsFileName = "",
  cohortPheno = "", phenText = "Test Pheno", order = 1,
  testText = "Phenotype", cohortDatabaseSchema = "",
  cohortTable = "", estPPV = 1, cdmShortName = "CDM")
}
\arguments{
\item{connectionDetails}{ConnectionDetails created using the function createConnectionDetails
in the DatabaseConnector package.}

\item{cutPoints}{A list of threshold predictions for the evaluations.  Include "EV" for
the expected value}

\item{resultsFileName}{The full file name with path for the evalution file}

\item{cohortPheno}{The number of the cohort of the phenotype algorithm to test}

\item{phenText}{A string to identify the phenotype algorithm in the outpuit file}

\item{order}{The order of this algorithm for sorting in the output file (used when
there are multiple phenotypes to test)}

\item{testText}{Descriptive name for the model}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the cohort
data used to define the at risk cohort is available. Requires read
permissions to this database.}

\item{cohortTable}{The tablename that contains the at risk cohort. The expectation is
cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{estPPV}{The positive predictive value estimate used in developing the model}

\item{cdmShortName}{A short name for the current database (CDM)}
}
\value{
A list containg 2 dataframes: 1) results - a dataframe with the results from the phenotype
algorithm evaluation 2) misses - a dataframe with a sample of subject ids for TPs, FPs, TNs, and FNs
for the 50 percent and over prediction threshold
}
\description{
Test phenotype algorithms
}
\details{
This function will perform the phenotype algorithm evaluation using the evaluation cohort returned
from createEvalCohort and the phenotype algorithm cohort specified
}
