% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateEvaluationCohortArgs}
\alias{createCreateEvaluationCohortArgs}
\title{Create a parameter object for the function createEvaluationCohort}
\usage{
createCreateEvaluationCohortArgs(
  xSpecCohortId,
  daysFromxSpec = 0,
  xSensCohortId = prevalenceCohortId,
  prevalenceCohortId,
  caseCohortId = prevalenceCohortId,
  caseFirstOccurrenceOnly = TRUE,
  xSpecCohortSize = 5000,
  covariateSettings = createDefaultCovariateSettings(excludedCovariateConceptIds = c(),
    addDescendantsToExclude = TRUE),
  modelPopulationCohortId = 0,
  modelPopulationCohortIdStartDay = 0,
  modelPopulationCohortIdEndDay = 0,
  inclusionEvaluationCohortId = 0,
  inclusionEvaluationDaysFromStart = 0,
  inclusionEvaluationDaysFromEnd = 0,
  exclusionEvaluationCohortId = 0,
  exclusionEvaluationDaysFromStart = 0,
  exclusionEvaluationDaysFromEnd = 0,
  minimumOffsetFromStart = 0,
  minimumOffsetFromEnd = 30,
  modelBaseSampleSize = 25000,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  visitLength = 0,
  visitType = c(9201, 9202, 9203, 581477, 262),
  gender = c(8507, 8532),
  race = 0,
  ethnicity = 0,
  startDate = "19001010",
  endDate = "21000101",
  falsePositiveNegativeSubjects = 10,
  modelId = "main",
  evaluationCohortId = "main",
  randomVisitTable = "",
  excludeModelFromEvaluation = FALSE,
  removeSubjectsWithFutureDates = TRUE,
  saveEvaluationCohortPlpData = FALSE
)
}
\arguments{
\item{xSpecCohortId}{The number of the "extremely specific (xSpec)"
cohortdefinition id in the cohort table (for noisy
positives).}

\item{daysFromxSpec}{Number of days from xSpec condition until analyzed visit}

\item{xSensCohortId}{The number of the "extremely sensitive (xSens)"
cohortdefinition id in the cohort table (for noisy
negatives).}

\item{prevalenceCohortId}{The number of the cohort definition id to determine
the disease prevalence.}

\item{caseCohortId}{The number of the cohort definition id to determine cases in the evaluation cohort}

\item{caseFirstOccurrenceOnly}{Set to true if only the first occurrence per subject in the case cohort is to be used}

\item{covariateSettings}{A covariateSettings object as generated
using createCovariateSettings().}

\item{modelPopulationCohortId}{The number of the cohort to be used as a base population for
the model. If set to 0, the entire database population will be
used.}

\item{modelPopulationCohortIdStartDay}{The number of days relative to the mainPopulationCohortId
cohort start date to begin including visits.}

\item{modelPopulationCohortIdEndDay}{The number of days relative to the mainPopulationCohortId
cohort start date to end including visits.}

\item{inclusionEvaluationCohortId}{The number of the cohort of the population to be used to designate which visits
are are eligible to be in the evaluation cohort}

\item{inclusionEvaluationDaysFromStart}{The number of days from the cohort start date of the inclusionEvaluationCohortId
to start eligible included visits}

\item{inclusionEvaluationDaysFromEnd}{The number of days from the cohort end date of the inclusionEvaluationCohortId
to end eligible included visits}

\item{exclusionEvaluationCohortId}{The number of the cohort of the population to be used to designate which visits
are NOT eligible to be in the evaluation cohort}

\item{exclusionEvaluationDaysFromStart}{The number of days from the cohort start date of the exclusionEvaluationCohortId
to start ineligible included visits}

\item{exclusionEvaluationDaysFromEnd}{The number of days from the cohort end date of the exclusionEvaluationCohortId
to end ineligible included visits}

\item{minimumOffsetFromStart}{Minimum number of days to offset for the analysis visit from the start of the observation period}

\item{minimumOffsetFromEnd}{Minimum number of days to offset for the analysis visit from the end of the observation period}

\item{modelBaseSampleSize}{The number of non-xSpec subjects to include in the model}

\item{baseSampleSize}{The maximum number of subjects in the evaluation cohort.}

\item{lowerAgeLimit}{The lower age for subjects in the model.}

\item{upperAgeLimit}{The upper age for subjects in the model.}

\item{visitLength}{The minimum length of index visit for acute outcomes.}

\item{visitType}{The concept_id for the visit type.}

\item{gender}{The gender(s) to be included.}

\item{race}{The race(s) to be included.}

\item{ethnicity}{The ethnicity(s) to be included.}

\item{startDate}{The starting date for including subjects in the model.}

\item{endDate}{The ending date for including subjects in the model.}

\item{falsePositiveNegativeSubjects}{Number of subjects to include for evaluating false positives and negatives}

\item{modelId}{A string used to generate the file names for this model.}

\item{evaluationCohortId}{A string used to generate the file names for this evaluation cohort.}

\item{randomVisitTable}{Table stored in work directory with pre-selected random visits in format of visit_occurrence table}

\item{excludeModelFromEvaluation}{Should subjects used in the model be excluded from the
evaluation cohort?}

\item{removeSubjectsWithFutureDates}{For buggy data with data in the future: ignore subjects
with dates in the future?}

\item{saveEvaluationCohortPlpData}{Should the large PLP file for the evaluation cohort be
saved? To be used for debugging purposes.}
}
\description{
Create a parameter object for the function createEvaluationCohort
}
\details{
Create an object defining the parameter values.
}
