% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateEvaluationCohortArgs}
\alias{createCreateEvaluationCohortArgs}
\title{Create a parameter object for the function createEvaluationCohort}
\usage{
createCreateEvaluationCohortArgs(
  xSpecCohortId,
  xSensCohortId,
  prevalenceCohortId,
  xSpecCohortSize = 5000,
  covariateSettings = createDefaultChronicCovariateSettings(excludedCovariateConceptIds
    = c(), addDescendantsToExclude = TRUE),
  modelPopulationCohortId = 0,
  modelPopulationCohortIdStartDay = 0,
  modelPopulationCohortIdEndDay = 0,
  evaluationPopulationCohortId = 0,
  evaluationPopulationCohortIdStartDay = 0,
  evaluationPopulationCohortIdEndDay = 0,
  modelBaseSampleSize = 15000,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  visitLength = 0,
  visitType = c(9201, 9202, 9203, 581477, 262),
  gender = c(8507, 8532),
  race = 0,
  ethnicity = 0,
  startDate = "19001010",
  endDate = "21000101",
  modelId = "main",
  evaluationCohortId = "main",
  excludeModelFromEvaluation = FALSE,
  removeSubjectsWithFutureDates = TRUE,
  saveEvaluationCohortPlpData = FALSE,
  modelType = "acute"
)
}
\arguments{
\item{xSpecCohortId}{The number of the "extremely specific (xSpec)"
cohortdefinition id in the cohort table (for noisy
positives).}

\item{xSensCohortId}{The number of the "extremely sensitive (xSens)"
cohortdefinition id in the cohort table (for noisy
negatives).}

\item{prevalenceCohortId}{The number of the cohort definition id to determine
the disease prevalence.}

\item{xSpecCohortSize}{The recommended xSpec sample size to use in model (default = 5000)}

\item{covariateSettings}{A covariateSettings object as generated
using createCovariateSettings().}

\item{modelPopulationCohortId}{The number of the cohort to be used as a base population for
the model. If set to 0, the entire database population will be
used.}

\item{modelPopulationCohortIdStartDay}{The number of days relative to the mainPopulationCohortId
cohort start date to begin including visits.}

\item{modelPopulationCohortIdEndDay}{The number of days relative to the mainPopulationCohortId
cohort start date to end including visits.}

\item{evaluationPopulationCohortId}{The number of the cohort to be used as a base population for
the evalution cohort. If set to 0, the entire database population will be
used.}

\item{evaluationPopulationCohortIdStartDay}{The number of days relative to the evaluationPopulationCohortId
cohort start date to begin including visits.}

\item{evaluationPopulationCohortIdEndDay}{The number of days relative to the evaluationPopulationCohortId
cohort start date to end including visits.}

\item{modelBaseSampleSize}{The number of non-xSpec subjects to include in the model}

\item{baseSampleSize}{The maximum number of subjects in the evaluation cohort.}

\item{lowerAgeLimit}{The lower age for subjects in the model.}

\item{upperAgeLimit}{The upper age for subjects in the model.}

\item{visitLength}{The minimum length of index visit for acute outcomes.}

\item{visitType}{The concept_id for the visit type.}

\item{gender}{The gender(s) to be included.}

\item{race}{The race(s) to be included.}

\item{ethnicity}{The ethnicity(s) to be included.}

\item{startDate}{The starting date for including subjects in the model.}

\item{endDate}{The ending date for including subjects in the model.}

\item{modelId}{A string used to generate the file names for this model.}

\item{evaluationCohortId}{A string used to generate the file names for this evaluation cohort.}

\item{excludeModelFromEvaluation}{Should subjects used in the model be excluded from the
evaluation cohort?}

\item{removeSubjectsWithFutureDates}{For buggy data with data in the future: ignore subjects
with dates in the future?}

\item{saveEvaluationCohortPlpData}{Should the large PLP file for the evaluation cohort be
saved? To be used for debugging purposes.}

\item{modelType}{The type of health outcome in the model either "acute"
or"chronic".}
}
\description{
Create a parameter object for the function createEvaluationCohort
}
\details{
Create an object defining the parameter values.
}
