% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createCreateEvaluationCohortArgs}
\alias{createCreateEvaluationCohortArgs}
\title{Create a parameter object for the function createEvaluationCohort}
\usage{
createCreateEvaluationCohortArgs(
  xSpecCohortId,
  xSensCohortId,
  prevalenceCohortId = xSensCohortId,
  xSpecCohortSize = NULL,
  covariateSettings = createDefaultChronicCovariateSettings(excludedCovariateConceptIds
    = c(), addDescendantsToExclude = FALSE),
  mainPopulationCohortId = 0,
  mainPopulationCohortIdStartDay = 0,
  mainPopulationCohortIdEndDay = 0,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  visitLength = 3,
  visitType = c(9201),
  gender = c(8507, 8532),
  startDate = "19001010",
  endDate = "21000101",
  excludeModelFromEvaluation = TRUE,
  removeSubjectsWithFutureDates = TRUE,
  saveEvaluationCohortPlpData = FALSE,
  modelType = "chronic"
)
}
\arguments{
\item{xSpecCohortId}{The number of the "extremely specific (xSpec)"
cohortdefinition id in the cohort table (for noisy
positives).}

\item{xSensCohortId}{The number of the "extremely sensitive (xSens)"
cohortdefinition id in the cohort table (for noisy
negatives).}

\item{prevalenceCohortId}{The number of the cohort definition id to determine
thedisease prevalence.}

\item{xSpecCohortSize}{The recommended xSpec sample size to use in model (default = NULL)}

\item{covariateSettings}{A covariateSettings object as generated
usingcreateCovariateSettings().}

\item{mainPopulationCohortId}{The number of the cohort to be used as a base population
forthe model. If set to 0, the entire database population
will beused.}

\item{mainPopulationCohortIdStartDay}{The number of days relative to the
mainPopulationCohortIdcohort start date to begin including
visits.}

\item{mainPopulationCohortIdEndDay}{The number of days relative to the
mainPopulationCohortIdcohort start date to end including
visits.}

\item{baseSampleSize}{The maximum number of subjects in the evaluation cohort.}

\item{lowerAgeLimit}{The lower age for subjects in the model.}

\item{upperAgeLimit}{The upper age for subjects in the model.}

\item{visitLength}{The minimum length of index visit for acute outcomes.}

\item{visitType}{The concept_id for the visit type.}

\item{gender}{The gender(s) to be included.}

\item{startDate}{The starting date for including subjects in the model.}

\item{endDate}{The ending date for including subjects in the model.}

\item{excludeModelFromEvaluation}{Should subjects used in the model be excluded from the
evaluation cohort?}

\item{removeSubjectsWithFutureDates}{For buggy data with data in the future: ignore subjects
withdates in the future?}

\item{saveEvaluationCohortPlpData}{Should the large PLP file for the evaluation cohort be
saved? To beused for debugging purposes.}

\item{modelType}{The type of health outcome in the model either "acute"
or"chronic".}
}
\description{
Create a parameter object for the function createEvaluationCohort
}
\details{
Create an object defining the parameter values.
}
