% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCovariateSettings.R
\name{createDefaultCovariateSettings}
\alias{createDefaultCovariateSettings}
\title{Create the default covariate settings}
\usage{
createDefaultCovariateSettings(
  excludedCovariateConceptIds = c(),
  includedCovariateIds = c(),
  includedCovariateConceptIds = c(),
  addDescendantsToExclude = FALSE,
  startDayWindow1 = 0,
  endDayWindow1 = 9999,
  startDayWindow2 = NULL,
  endDayWindow2 = NULL,
  startDayWindow3 = NULL,
  endDayWindow3 = NULL
)
}
\arguments{
\item{excludedCovariateConceptIds}{A list of conceptIds to exclude from featureExtraction.  These
should include all concept_ids that were used to define the
xSpec model (default=NULL)}

\item{includedCovariateIds}{A list of covariate IDs that should be restricted to.}

\item{includedCovariateConceptIds}{A list of covariate concept IDs that should be restricted to.}

\item{addDescendantsToExclude}{Should descendants of excluded concepts also be excluded?
(default=FALSE)}

\item{startDayWindow1}{The day to start time window 1 for feature extraction}

\item{endDayWindow1}{The day to end time window 1 for feature extraction}

\item{startDayWindow2}{The day to start time window 2 for feature extraction}

\item{endDayWindow2}{The day to end time window 2 for feature extraction}

\item{startDayWindow3}{The day to start time window 3 for feature extraction}

\item{endDayWindow3}{The day to end time window 3 for feature extraction #'}
}
\description{
Create the default covariate settings for acute or chronic diseases
}
\details{
Function to create the default covariate settings for acute or chronic diseases
}
\examples{
covSettings <- createDefaultCovariateSettings(
  excludedCovariateConceptIds = c(201826, 44786627),
  addDescendantsToExclude = TRUE
)

}
