% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEvalCohort.R
\name{createEvaluationCohort}
\alias{createEvaluationCohort}
\title{Create the evaluation cohort}
\usage{
createEvaluationCohort(
  connectionDetails,
  xSpecCohort,
  xSensCohort,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortDatabaseTable,
  outDatabaseSchema,
  evaluationOutputFileName,
  modelOutputFileName,
  mainPopulationCohort = 0,
  mainPopulationCohortStartDay = 0,
  mainPopulationCohortEndDay = 0,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  startDays = 0,
  endDays = 10000,
  visitLength = 3,
  gender = c(8507, 8532),
  startDate = "19001010",
  endDate = "21000101",
  cdmVersion = "5",
  outFolder = getwd(),
  excludeModelFromEvaluation = TRUE,
  savePlpData = FALSE,
  modelType = "chronic"
)
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function createConnectionDetails
in the DatabaseConnector package}

\item{xSpecCohort}{The number of the "extremely specific (xSpec)" cohort definition id in
the cohort table (for noisy positives)}

\item{xSensCohort}{The number of the "extremely sensitive (xSens)" cohort definition id in
the cohort table (for noisy negatives)}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this database. On SQL Server, this should
specifiy both the database and the schema, so for example
'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the cohort
data used to define the at risk cohort is available. Requires read
permissions to this database.}

\item{cohortDatabaseTable}{The tablename that contains the at risk cohort. The expectation is
cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outDatabaseSchema}{The name of the database schema that is the location where the data
used to define the outcome cohorts is available. Requires read
permissions to this database.}

\item{evaluationOutputFileName}{A string designation for the evaluation cohort file}

\item{modelOutputFileName}{A string designation for the training model file}

\item{mainPopulationCohort}{The number of the cohort to be used as a base population for the model
(default=NULL)}

\item{mainPopulationCohortStartDay}{The number of days relative to the mainPopulationCohort cohort start date
to begin including visits (default=0)}

\item{mainPopulationCohortEndDay}{The number of days relative to the mainPopulationCohort cohort start date
to end including visits (default=0)}

\item{baseSampleSize}{The maximum number of subjects in the evaluation cohort (default=2M)}

\item{lowerAgeLimit}{The lower age for subjects in the model (default=NULL)}

\item{upperAgeLimit}{The upper age for subjects in the model (default=NULL)}

\item{startDays}{The days to include prior to the cohort start date (default=0)}

\item{endDays}{The days to include after the cohort start date (default=10000)}

\item{visitLength}{The minimum length of index visit (default=3)}

\item{gender}{The gender(s) to be included (default c(8507, 8532))}

\item{startDate}{The starting date for including subjects in the model (default=NULL)}

\item{endDate}{The ending date for including subjects in the model (default=NULL)}

\item{cdmVersion}{The CDM version of the database (default=5)}

\item{outFolder}{The folder where the output files will be written (default=working directory)}

\item{excludeModelFromEvaluation}{Should subjects used in the model be excluded from the evaluation cohort (default = TRUE)}

\item{savePlpData}{Should large PLP data file be saved (default=FALSE)}

\item{modelType}{The type of health outcome in the model either "acute" or "chronic" (Default = "chronic")}
}
\description{
Create the evaluation cohort
}
\details{
Creates the evaluation cohort and applies a diagnostic prediction model for determination of a
probability for the health outcome of interest
}
