% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEvalCohort.R
\name{createEvaluationCohort}
\alias{createEvaluationCohort}
\title{Create the evaluation cohort}
\usage{
createEvaluationCohort(connectionDetails, xSpecCohort, cdmDatabaseSchema,
  cohortDatabaseSchema, cohortDatabaseTable, outDatabaseSchema,
  evaluationOutputFileName, modelOutputFileName,
  mainPopulationCohort = 0, lowerAgeLimit = 0, upperAgeLimit = 120,
  startDays = -10000, endDays = 10000, gender = c(8507, 8532),
  startDate = "19001010", endDate = "21000101", cdmVersion = "5",
  outFolder = getwd())
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function createConnectionDetails
in the DatabaseConnector package.}

\item{xSpecCohort}{The number of the "extremely specific (xSpec)" cohort definition id in
the cohort table (for noisy positives)}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance.
Requires read permissions to this database. On SQL Server, this should
specifiy both the database and the schema, so for example
'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where the cohort
data used to define the at risk cohort is available. Requires read
permissions to this database.}

\item{cohortDatabaseTable}{The tablename that contains the at risk cohort. The expectation is
cohortTable has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{outDatabaseSchema}{The name of the database schema that is the location where the data
used to define the outcome cohorts is available. Requires read
permissions to this database.}

\item{evaluationOutputFileName}{A string designation for the evaluation cohort file}

\item{modelOutputFileName}{A string designation for the training model file}

\item{mainPopulationCohort}{The number of the cohort to be used as a base population for the model
(default=NULL)}

\item{lowerAgeLimit}{The lower age for subjects in the model (default=NULL)}

\item{upperAgeLimit}{The upper age for subjects in the model (default=NULL)}

\item{startDays}{The days to include prior to the cohort start date (default=-10000)}

\item{endDays}{The days to include after the cohort start date (default=10000)}

\item{gender}{The gender(s) to be included (default c(8507, 8532))}

\item{startDate}{The starting date for including subjects in the model (default=NULL)}

\item{endDate}{The ending date for including subjects in the model (default=NULL)}

\item{cdmVersion}{The CDM version of the database (default=5)}

\item{outFolder}{The folder where the output files will be written (default=working directory)}
}
\description{
Create the evaluation cohort
}
\details{
Creates the evaluation cohort and applies a diagnostic prediction model for determination of a
probability for the health outcome of interest
}
