% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEvaluationCohort.R
\name{createEvaluationCohort}
\alias{createEvaluationCohort}
\title{Create the evaluation cohort}
\usage{
createEvaluationCohort(
  connectionDetails,
  oracleTempSchema = NULL,
  xSpecCohortId,
  xSensCohortId,
  prevalenceCohortId = xSensCohortId,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  workDatabaseSchema,
  covariateSettings = createDefaultChronicCovariateSettings(excludedCovariateConceptIds
    = c(), addDescendantsToExclude = FALSE),
  mainPopulationCohortId = 0,
  mainPopulationCohortIdStartDay = 0,
  mainPopulationCohortIdEndDay = 0,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  visitLength = 3,
  visitType = c(9201),
  gender = c(8507, 8532),
  startDate = "19001010",
  endDate = "21000101",
  cdmVersion = "5",
  outFolder = getwd(),
  evaluationCohortId = "main",
  removeSubjectsWithFutureDates = TRUE,
  saveEvaluationCohortPlpData = FALSE,
  modelType = "chronic"
)
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function
createConnectionDetails in the DatabaseConnector package.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{xSpecCohortId}{The number of the "extremely specific (xSpec)" cohort
definition id in the cohort table (for noisy positives).}

\item{xSensCohortId}{The number of the "extremely sensitive (xSens)" cohort
definition id in the cohort table (for noisy negatives).}

\item{prevalenceCohortId}{The number of the cohort definition id to determine the
disease prevalence.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where
the cohort data used to define the at risk cohort is
available. Requires read permissions to this database.}

\item{cohortTable}{The tablename that contains the at risk cohort. The
expectation is cohortTable has format of COHORT table:
cohort_concept_id, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{workDatabaseSchema}{The name of the database schema that is the location where
a table can be created and afterwards removed.
Requires write permissions to this database.}

\item{covariateSettings}{A covariateSettings object as generated using
createCovariateSettings().}

\item{mainPopulationCohortId}{The number of the cohort to be used as a base population for
the model. If set to 0, the entire database population will be
used.}

\item{mainPopulationCohortIdStartDay}{The number of days relative to the mainPopulationCohortId
cohort start date to begin including visits.}

\item{mainPopulationCohortIdEndDay}{The number of days relative to the mainPopulationCohortId
cohort start date to end including visits.}

\item{baseSampleSize}{The maximum number of subjects in the evaluation cohort.}

\item{lowerAgeLimit}{The lower age for subjects in the model.}

\item{upperAgeLimit}{The upper age for subjects in the model.}

\item{visitLength}{The minimum length of index visit for acute outcomes.}

\item{visitType}{The concept_id for the visit type (default = 9201)}

\item{gender}{The gender(s) to be included.}

\item{startDate}{The starting date for including subjects in the model.}

\item{endDate}{The ending date for including subjects in the model.}

\item{cdmVersion}{The CDM version of the database.}

\item{outFolder}{The folder where the output files will be written.}

\item{evaluationCohortId}{A string used to generate the file names for this evaluation cohort.}

\item{removeSubjectsWithFutureDates}{For buggy data with data in the future: ignore subjects with
dates in the future?}

\item{saveEvaluationCohortPlpData}{Should the large PLP file for the evaluation cohort be saved? To be
used for debugging purposes.}

\item{modelType}{The type of health outcome in the model either "acute" or
"chronic".}
}
\description{
Create the evaluation cohort
}
\details{
Fits a diagnostic pretiction model, and uses it to create an evaluation cohort with
probabilities for the health outcome of interest.
}
