% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEvaluationCohort.R
\name{createEvaluationCohort}
\alias{createEvaluationCohort}
\title{Create the evaluation cohort}
\usage{
createEvaluationCohort(
  connectionDetails,
  oracleTempSchema = NULL,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  xSpecCohortId,
  daysFromxSpec = 0,
  xSensCohortId,
  prevalenceCohortId,
  xSpecCohortSize = 5000,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  workDatabaseSchema,
  covariateSettings = createDefaultCovariateSettings(excludedCovariateConceptIds = c(),
    addDescendantsToExclude = TRUE),
  modelPopulationCohortId = 0,
  modelPopulationCohortIdStartDay = 0,
  modelPopulationCohortIdEndDay = 0,
  evaluationPopulationCohortId = 0,
  evaluationPopulationCohortIdStartDay = 0,
  evaluationPopulationCohortIdEndDay = 0,
  modelBaseSampleSize = 15000,
  baseSampleSize = 2e+06,
  lowerAgeLimit = 0,
  upperAgeLimit = 120,
  visitLength = 0,
  visitType = c(9201, 9202, 9203, 262, 581477),
  gender = c(8507, 8532),
  race = 0,
  ethnicity = 0,
  startDate = "19001010",
  endDate = "21000101",
  cdmVersion = "5",
  outFolder = getwd(),
  modelId = "main",
  evaluationCohortId = "main",
  excludeModelFromEvaluation = FALSE,
  removeSubjectsWithFutureDates = TRUE,
  saveEvaluationCohortPlpData = FALSE
)
}
\arguments{
\item{connectionDetails}{connectionDetails created using the function
createConnectionDetails in the DatabaseConnector package.}

\item{oracleTempSchema}{DEPRECATED: use \code{tempEmulationSchema} instead.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{xSpecCohortId}{The number of the "extremely specific (xSpec)" cohort
definition id in the cohort table (for noisy positives).}

\item{daysFromxSpec}{Number of days from xSpec condition until analyzed visit}

\item{xSensCohortId}{The number of the "extremely sensitive (xSens)" cohort
definition id in the cohort table (for noisy negatives).}

\item{prevalenceCohortId}{The number of the cohort definition id to determine the
disease prevalence.}

\item{xSpecCohortSize}{The recommended xSpec sample size to use in model (default = NULL)}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where
the cohort data used to define the at risk cohort is
available. Requires read permissions to this database.}

\item{cohortTable}{The tablename that contains the at risk cohort. The
expectation is cohortTable has format of COHORT table:
cohort_concept_id, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{workDatabaseSchema}{The name of the database schema that is the location where
a table can be created and afterwards removed.
Requires write permissions to this database.}

\item{covariateSettings}{A covariateSettings object as generated using
createCovariateSettings().}

\item{modelPopulationCohortId}{The number of the cohort to be used as a base population for
the model. If set to 0, the entire database population will be
used.}

\item{modelPopulationCohortIdStartDay}{The number of days relative to the mainPopulationCohortId
cohort start date to begin including visits.}

\item{modelPopulationCohortIdEndDay}{The number of days relative to the mainPopulationCohortId
cohort start date to end including visits.}

\item{evaluationPopulationCohortId}{The number of the cohort to be used as a base population for
the evalution cohort. If set to 0, the entire database population will be
used.}

\item{evaluationPopulationCohortIdStartDay}{The number of days relative to the evaluationPopulationCohortId
cohort start date to begin including visits.}

\item{evaluationPopulationCohortIdEndDay}{The number of days relative to the evaluationPopulationCohortId
cohort start date to end including visits.}

\item{modelBaseSampleSize}{The number of non-xSpec subjects to include in the model}

\item{baseSampleSize}{The maximum number of subjects in the evaluation cohort.}

\item{lowerAgeLimit}{The lower age for subjects in the model.}

\item{upperAgeLimit}{The upper age for subjects in the model.}

\item{visitLength}{The minimum length of index visit for acute outcomes.}

\item{visitType}{The concept_id for the visit type.}

\item{gender}{The gender(s) to be included.}

\item{race}{The race(s) to be included.}

\item{ethnicity}{The ethnicity(s) to be included.}

\item{startDate}{The starting date for including subjects in the model.}

\item{endDate}{The ending date for including subjects in the model.}

\item{cdmVersion}{The CDM version of the database.}

\item{outFolder}{The folder where the output files will be written.}

\item{modelId}{A string used to generate the file names for this model.}

\item{evaluationCohortId}{A string used to generate the file names for this evaluation cohort.}

\item{excludeModelFromEvaluation}{Should subjects used in the model be excluded from the evaluation cohort?}

\item{removeSubjectsWithFutureDates}{For buggy data with data in the future: ignore subjects with
dates in the future?}

\item{saveEvaluationCohortPlpData}{Should the large PLP file for the evaluation cohort be saved? To be
used for debugging purposes.}
}
\description{
Create the evaluation cohort
}
\details{
Fits a diagnostic prediction model, and uses it to create an evaluation cohort with
probabilities for the health outcome of interest.
}
