% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createPhevaluatorAnalysis}
\alias{createPhevaluatorAnalysis}
\title{Create a PheValuator analysis specification}
\usage{
createPhevaluatorAnalysis(
  analysisId,
  description,
  createEvaluationCohortArgs,
  testPhenotypeAlgorithmArgs
)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific
set of analysis choices.}

\item{description}{A short description of the analysis.}

\item{createEvaluationCohortArgs}{An object representing the arguments to be used when calling
the \code{\link{createCreateEvaluationCohortArgs}} function.}

\item{testPhenotypeAlgorithmArgs}{Should the \code{\link{createTestPhenotypeAlgorithmArgs}} function be used in this
analysis?}
}
\description{
Create a PheValuator analysis specification
}
\details{
Create a set of analysis choices, to be used with the \code{\link{runPheValuatorAnalyses}} function.
}
