% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createTestPhenotypeAlgorithmArgs}
\alias{createTestPhenotypeAlgorithmArgs}
\title{Create a parameter object for the function testPhenotypeAlgorithm}
\usage{
createTestPhenotypeAlgorithmArgs(
  cutPoints = c("EV"),
  phenotypeCohortId,
  washoutPeriod = 0,
  splayPrior = 7,
  splayPost = 7
)
}
\arguments{
\item{cutPoints}{A list of threshold predictions for the evaluations.  Include "EV"for the
expected value}

\item{phenotypeCohortId}{The ID of the cohort to evaluate in the specified cohort table.}

\item{washoutPeriod}{The mininum required continuous observation time prior to indexdate for
subjects within the cohort to test (Default = 0).}

\item{splayPrior}{The number of days to allow for test phenotype visit date prior to evaluation date}

\item{splayPost}{The number of days to allow for test phenotype visit date after evaluation date}
}
\description{
Create a parameter object for the function testPhenotypeAlgorithm
}
\details{
Create an object defining the parameter values.
}
