% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createTestPhenotypeAlgorithmArgs}
\alias{createTestPhenotypeAlgorithmArgs}
\title{Create a parameter object for the function testPhenotypeAlgorithm}
\usage{
createTestPhenotypeAlgorithmArgs(
  cutPoints = c(0.1, 0.2, 0.3, 0.4, 0.5, "EV", 0.6, 0.7, 0.8, 0.9),
  phenotypeCohortId,
  washoutPeriod = 0
)
}
\arguments{
\item{cutPoints}{A list of threshold predictions for the evaluations.  Include "EV"for the
expected value}

\item{phenotypeCohortId}{The ID of the cohort to evaluate in the specified cohort table.}

\item{washoutPeriod}{The mininum required continuous observation time prior to indexdate for
subjects within the cohort to test (Default = 0).}
}
\description{
Create a parameter object for the function testPhenotypeAlgorithm
}
\details{
Create an object defining the parameter values.
}
