% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadPheValuatorAnalysisList}
\alias{loadPheValuatorAnalysisList}
\title{Load a list of phevaluatorAnalysis from file}
\usage{
loadPheValuatorAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{pheValuatorAnalysis}.
}
\description{
Load a list of objects of type \code{pheValuatorAnalysis} from file. The file is in JSON format.
}
