% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadPhevaluatorAnalysisList}
\alias{loadPhevaluatorAnalysisList}
\title{Load a list of phevaluatorAnalysis from file}
\usage{
loadPhevaluatorAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{phevaluatorAnalysis}.
}
\description{
Load a list of objects of type \code{phevaluatorAnalysis} from file. The file is in JSON format.
}
