% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runPheValuatorAnalyses}
\alias{runPheValuatorAnalyses}
\title{Run a list of analyses}
\usage{
runPheValuatorAnalyses(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  workDatabaseSchema = cdmDatabaseSchema,
  cdmVersion = 5,
  outputFolder,
  pheValuatorAnalysisList
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance. Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{cohortDatabaseSchema}{The name of the database schema that is the location where
the cohort data used to define the at risk cohort is
available. Requires read permissions to this database.}

\item{cohortTable}{The tablename that contains the at risk cohort. The
expectation is cohortTable has format of COHORT table:
cohort_concept_id, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{workDatabaseSchema}{The name of the database schema that is the location where
a table can be created and afterwards removed.
Requires write permissions to this database.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
"5".}

\item{outputFolder}{Name of the folder where all the outputs will written to.}

\item{pheValuatorAnalysisList}{A list of objects of type \code{pheValuatorAnalysis} as created using
the \code{\link{createPheValuatorAnalysis}} function.}
}
\value{
A data frame specifiying where the constructed evaluation cohort and phenotype evaluation results can be found
in the local file system.
}
\description{
Run a list of analyses
}
\details{
Run a list of analyses.
}
